requirements.txt
pyttsx3 
pyaudio
perplexityai
PerplexiPy
python-dotenv